% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_checks.R
\name{clean_data_renamer}
\alias{clean_data_renamer}
\title{rename the clean data}
\usage{
clean_data_renamer()
}
\arguments{
\item{data}{Dataframe with "start" and "end" columns.}

\item{duration_threshold_lower}{Minimum number of minutes to complete the form}

\item{duration_threshold_upper}{Maximum number of minutes to complete the form}
}
\value{
A dataframe with one row per potential issue. It has columns for the corresponding row index in the original data; the suspicious value (survey that is longer or shorter than it should be); the variable name in the original dataset in which the suspicious value occured; A description of the issue type.
}
\description{
Rename the clean data
}
